refac: move PrefectureSelector component files to PrefectureSelector folder
